feat: implement creating invoices automatically
This saves you manually doing these steps in Harvest, as invoices are "automatically" generated on the Toptal side once a timesheet is submitted. Unfortunately Toptal doesn't have an API to upload time information & submit it. There is a caveat to the current implementation here: It DOES NOT check if you've already sent an invoice for the given period, that's on you.